Several const-correctness and warnings fixes.  (If you have time, it would probably be good if someone would look over these changes...  I have reviewed them, but the more eyes the better.)

+++ cal3d/src/cal3d/hardwaremodel.h
@@ -128,7 +128,7 @@ private:
 	
 		Cal::UserData getMapUserData(int mapId);
 		
-		bool selectHardwareMesh(int meshId);  
+		bool selectHardwareMesh(size_t meshId);
 		
 	private:
 		bool canAddFace(CalHardwareMesh &hardwareMesh, CalCoreSubmesh::Face & face,std::vector<CalCoreSubmesh::Vertex>& vectorVertex, int maxBonesPerMesh);
